Postpone Maskables spawning to work around the Maskable-after-Graphic issue
See this post for some information:
https://forum.unity.com/threads/soft-mask-a-replacement-of-unity-ui-mask-with-alpha-support.454707/page-7#post-6348642

It's a n old Unity 'bug', but now I'm starting to think that it's
actually not a bug... probably I should not put my Maskables
before any Graphic?

This commit doesn't do this directly, it mainly aimed at the
specific issue when using TextMesh Pro with several fonts.