Make it possible to have more than one SoftMaskable on a GameObject
It's required for case when the same object falls under a mask
and leaves it back several times in a frame. Before the previous
commit is wasn't needed because SoftMaskable was destroyed
with DestroyImmediate. Now we postpond destroy to the end of
frame, so, there may be more than one SoftMaskable on an
object at the moment.